Remove duplicate S3 test helper functions
