+++ mountpoint-s3/tests/fuse_tests/fork_test.rs
@@ -10,11 +10,10 @@ use std::process::Stdio;
 use std::{path::PathBuf, process::Command};
 use test_case::test_case;
 
-use crate::common::fuse::{
-    create_objects, get_test_bucket_and_prefix, get_test_bucket_forbidden, get_test_region, read_dir_to_entry_names,
-};
+use crate::common::fuse::read_dir_to_entry_names;
+use crate::common::s3::{create_objects, get_test_bucket_and_prefix, get_test_bucket_forbidden, get_test_region};
 #[cfg(not(feature = "s3express_tests"))]
-use crate::common::fuse::{get_subsession_iam_role, tokio_block_on};
+use crate::common::s3::{get_subsession_iam_role, tokio_block_on};
 
 const MAX_WAIT_DURATION: std::time::Duration = std::time::Duration::from_secs(10);
